Fixing Issue #9
When an sqlQuery is started before a profiling node, there was no node to attach the sqlQuery to.  To alleviate this problem if there is no node available, I create a default top level node named "Profiler Default Top Level" and attach the sqlQuery to that.  This might create unexpected results if nodes are started later on, as they'd be grouped underneath the default top level node.